Increased kill time, was getting false kills

+++ src/wordslayer.rb
@@ -5,7 +5,7 @@ require 'fileutils'
 
 
 def get_process_array(wmi)
-    processes=wmi.ExecQuery("select * from win32_process where name='WINWORD.EXE' or name='DW20.EXE'")
+    processes=wmi.ExecQuery("select * from win32_process where name='WINWORD.EXE'")
     ary=[]
     processes.each {|p|
         ary << p.ProcessId
@@ -15,11 +15,10 @@ def get_process_array(wmi)
 end
 
 def delete_temp_files
-    tempfiles='C:/Documents and Settings/Administrator/Local Settings/Temporary Internet Files/Content.Word/*WR*.tmp'
-    fuzzfiles='C:/fuzzclient/~$*.doc'
-
-    [tempfiles,fuzzfiles].each {|pattern|
-        Dir.glob(pattern, File::FNM_DOTMATCH).each {|fn| 
+        patterns=['R:/Temp/**/*.*', 'R:/Temporary Internet Files/**/*.*', 'R:/fuzzclient/~$*.doc']
+        patterns.each {|pattern|
+        Dir.glob(pattern, File::FNM_DOTMATCH).each {|fn|
+            next if File.directory?(fn)
             begin
                 FileUtils.rm_f(fn)
             rescue
@@ -27,32 +26,30 @@ def delete_temp_files
             end
             print "@";$stdout.flush
         }
-    }
+        }
 end
 
 word_instances=Hash.new(0)
+wmi = WIN32OLE.connect("winmgmts://")
+FileUtils.mkdir_p 'R:/Temp'
 begin
-    wmi = WIN32OLE.connect("winmgmts://")
     loop do
         procs=get_process_array(wmi)
         word_instances.delete_if {|pid,kill_level| not procs.include?(pid)}
         procs.each {|p| word_instances[p]+=1}
         word_instances.each {|pid,kill_level|
-            if kill_level > 8
+            if kill_level > 1 # seen before, try and kill
                 Process.kill(9,pid)
-                print "<!#{pid}!>";$stdout.flush
-            elsif kill_level > 1 # seen before, try and kill
-                Process.kill(1,pid)
                 print "<#{pid}>";$stdout.flush
                 word_instances[pid]=9
             end
         }
         print '*';$stdout.flush
-        sleep(10)
+        sleep(6)
         delete_temp_files
     end
 rescue
     puts "Wordslayer: PK: #{$!}"
-    sleep(5)
+    sleep(1)
     retry
 end
